ctmm FAQ

Frequently asked questions for the ctmm package.

Details

General recommendations:

Do not save workspaces between sessions. They can become corrupted.

Upgrade R to the latest version and update all of your packages.

Work through the vignettes vignette("variogram") and vignette("akde").

Check the http://www2.physics.umd.edu/~hfleming/ to see if there is a beta release that fixes any bugs.

Check the https://groups.google.com/forum/#!forum/ctmm-user.

Specific issues:

Database corruption, missing functions, and other weird errors

R likely did not install the package correctly. I.e., some files may have failed to overwrite previous versions. Uninstall ctmm, restart R, and install ctmm again.

manipulate::isAvailable is not found

You probably have an outdated copy of the manipulate package installed. Update R to the latest version and then update all of your packages. This seems to happen frequently with the MacOS release of R.

The manipulate panel does not popup in variogram.fit or zoom

Click the gear icon in the upper-left corner of the plot window.

There is no save button in variogram.fit

Maximize the plot window and/or increase your screen resolution.

plot complains about the datatype or has weird errors

Namespace collision sometimes occurs from raster or sp against move and ctmm. Restart R and only load the ctmm package.

Infinite recursion error

ctmm has no recursive functions, so I am not exactly sure what causes this error, but it only occurs with certain versions of R on certain computer architectures. There are several solutions that have worked for people, including restarting R in a fresh session and updating their software. Alternatively:

Reboot your computer.

Increase the allowed number of nested expressions within R via options(expressions=10000) or some other large number.

Try a different computer.

ctmm.select only returns one model with verbose=TRUE

The newer version of ctmm.select is much more intelligent, necessitated by the exploding number of new model permutations. Any simpler models were judged to be so unlikely that they were not attempted. You can set level=1 to force alternative model fitting, but the additional models should lose badly in model selection (please email me if I am wrong).