Parametric bootstrap continuous-time movement models

This function allows the point estimates and confidence intervals of an initial estimated movement model to be improved by parametric boostrap, as described in Fleming et al (2019).

ctmm.boot(data,CTMM,method=CTMM$method,AICc=FALSE,iterate=FALSE,robust=FALSE,error=0.01,
          cores=1,trace=TRUE,...)

Arguments

data

Timeseries data represented as a telemetry object.

CTMM

A ctmm movement-model object from the output of ctmm.fit containing the initial parameter estimates.

method

Fitting method to use: "ML", "HREML", "pREML", "pHREML", or "REML". See ctmm.fit for descriptions.

AICc

Run dual set of simulations to approximate AICc values via Kullback–Leibler divergence. Otherwise, only the AIC is updated.

iterate

Iteratively solve for the parameters such that the average estimate (of method) is that of the data, whereas with iterate=FALSE only the first-order correction is calculated from the initial estimate.

robust

Uses robust estimates of the average and covariation for debiasing. Useful when parameters are near boundaries.

error

Relative standard error target for bootstrap ensemble estimates and nonlinear iterations.

cores

Number of simulations to run in parallel. cores=NULL will use all cores, while cores<0 will reserve abs(cores).

trace

Report progress updates. Can be among 0:2 with increasing detail.

...

Further arguments passed to ctmm.fit.

Value

A model fit object with relatively unbiased estimates of location covariance, and autocorrelation timescales (and more accurate CIs than ctmm.fit). If AICc=TRUE, then, in addition to an updated AICc slot, the model fit object will also contain a VAR.AICc slot quantifying the numerical variance in the AICc estimate. This variance can be decreased by decreasing argument error.

References

C. H. Fleming, M. J. Noonan, E. P. Medici, J. M. Calabrese, ``Overcoming the challenge of small effective sample sizes in home-range estimation'', Methods in Ecology and Evolution 10:10, 1679-1689 (2019) doi:10.1111/2041-210X.13270 .

Author

C. H. Fleming.

See also

Examples

# \donttest{
# Load package and data
library(ctmm)
data(gazelle)
DATA <- gazelle[[3]]

GUESS <- ctmm.guess(DATA,interactive=FALSE)
FIT <- ctmm.select(DATA,GUESS)

# some human-readable information
summary(FIT)
#> $name
#> [1] "OUf"
#> 
#> $DOF
#>      mean      area diffusion     speed 
#>  3.719081  4.105042  9.584254 18.886592 
#> 
#> $CI
#>                                           low         est        high
#> area (square kilometers)          7031.904969 25246.63797 54889.38775
#> τ (days)                             2.805659     4.43780     7.01941
#> speed (kilometers/day)               9.108161    11.74848    14.38350
#> diffusion (square kilometers/day)  104.656588   222.37629   383.70595
#> 

# in general, you will want to set iterate=TRUE
BOOT <- ctmm.boot(DATA,FIT,iterate=FALSE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |========                                                              |  12%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |=================                                                     |  25%  |                                                                              |==========================                                            |  37%  |                                                                              |=====================                                                 |  31%  |                                                                              |=======================                                               |  33%  |                                                                              |=========================                                             |  35%  |                                                                              |==========================                                            |  36%  |                                                                              |========================                                              |  35%  |                                                                              |=========================                                             |  35%  |                                                                              |========================                                              |  34%  |                                                                              |========================                                              |  35%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  35%  |                                                                              |=============================                                         |  41%  |                                                                              |===========================                                           |  39%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |=================================                                     |  48%  |                                                                              |===============================                                       |  44%  |                                                                              |===================================                                   |  50%  |                                                                              |=====================================                                 |  53%  |                                                                              |=======================================                               |  56%  |                                                                              |=======================================                               |  55%  |                                                                              |=====================================                                 |  53%  |                                                                              |==========================================                            |  60%  |                                                                              |============================================                          |  63%  |                                                                              |==============================================                        |  66%  |                                                                              |=================================================                     |  70%  |                                                                              |===========================================                           |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |=============================================                         |  64%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |===================================================                   |  73%  |                                                                              |=====================================================                 |  76%  |                                                                              |=====================================================                 |  75%  |                                                                              |==================================================                    |  71%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |==================================================                    |  71%  |                                                                              |===================================================                   |  73%  |                                                                              |======================================================                |  77%  |                                                                              |====================================================                  |  74%  |                                                                              |===================================================                   |  73%  |                                                                              |===========================================================           |  84%  |                                                                              |========================================================              |  80%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=======================================================               |  79%  |                                                                              |=========================================================             |  81%  |                                                                              |============================================================          |  85%  |                                                                              |==========================================================            |  83%  |                                                                              |=============================================================         |  87%  |                                                                              |=========================================================             |  82%  |                                                                              |===========================================================           |  85%  |                                                                              |=============================================================         |  87%  |                                                                              |===================================================================   |  95%  |                                                                              |====================================================================  |  98%  |                                                                              |==================================================================    |  95%  |                                                                              |==================================================================    |  94%  |                                                                              |=========================                                             |  36%  |                                                                              |=================================                                     |  47%  |                                                                              |=====================================                                 |  53%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |=======================================                               |  55%  |                                                                              |=====================================                                 |  53%  |                                                                              |=====================================                                 |  52%  |                                                                              |======================================                                |  54%  |                                                                              |=======================================                               |  55%  |                                                                              |======================================                                |  54%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  58%  |                                                                              |=========================================                             |  58%  |                                                                              |========================================                              |  58%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  63%  |                                                                              |==============================================                        |  65%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  68%  |                                                                              |===============================================                       |  68%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |===============================================                       |  68%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |=================================================                     |  69%  |                                                                              |==================================================                    |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  72%  |                                                                              |==================================================                    |  71%  |                                                                              |===================================================                   |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |=======================================================               |  79%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |=========================================================             |  81%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  82%  |                                                                              |=========================================================             |  81%  |                                                                              |==========================================================            |  82%  |                                                                              |=========================================================             |  81%  |                                                                              |========================================================              |  80%  |                                                                              |=======================================================               |  78%  |                                                                              |======================================================                |  78%  |                                                                              |========================================================              |  81%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  86%  |                                                                              |============================================================          |  86%  |                                                                              |============================================================          |  85%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  85%  |                                                                              |==========================================================            |  83%  |                                                                              |=========================================================             |  82%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |=========================================================             |  82%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  86%  |                                                                              |============================================================          |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  86%  |                                                                              |============================================================          |  86%  |                                                                              |===========================================================           |  84%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  88%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  89%  |                                                                              |==============================================================        |  88%  |                                                                              |===============================================================       |  90%  |                                                                              |==============================================================        |  89%  |                                                                              |==============================================================        |  88%  |                                                                              |=============================================================         |  88%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |===============================================================       |  89%  |                                                                              |==============================================================        |  89%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |==============================================================        |  88%  |                                                                              |=============================================================         |  87%  |                                                                              |=============================================================         |  88%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |================================================================      |  91%  |                                                                              |=================================================================     |  92%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |==================================================================    |  94%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  92%  |                                                                              |==================================================================    |  95%  |                                                                              |==================================================================    |  94%  |                                                                              |====================================================================  |  97%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |======================================================================|  99%  |                                                                              |======================================================================| 100%  |                                                                              |======================================================================|  99%  |                                                                              |===================================================================== |  98%  |                                                                              |====================================================================  |  97%  |                                                                              |===================================================================== |  98%  |                                                                              |====================================================================  |  98%  |                                                                              |====================================================================  |  97%  |                                                                              |==============================                                        |  43%  |                                                                              |====================================                                  |  51%  |                                                                              |======================================                                |  54%  |                                                                              |=====================================                                 |  54%  |                                                                              |=======================================                               |  55%  |                                                                              |========================================                              |  57%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |=========================================                             |  59%  |                                                                              |=========================================                             |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |==============================================                        |  66%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |===============================================                       |  67%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |================================================                      |  68%  |                                                                              |=================================================                     |  71%  |                                                                              |=================================================                     |  70%  |                                                                              |================================================                      |  69%  |                                                                              |====================================================                  |  75%  |                                                                              |=======================================================               |  78%  |                                                                              |=====================================================                 |  76%  |                                                                              |=========================================================             |  81%  |                                                                              |==========================================================            |  83%  |                                                                              |=======================================================               |  79%  |                                                                              |=========================================================             |  81%  |                                                                              |==========================================================            |  83%  |                                                                              |=========================================================             |  81%  |                                                                              |=======================================================               |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |=========================================================             |  81%  |                                                                              |==========================================================            |  83%  |                                                                              |============================================================          |  85%  |                                                                              |=============================================================         |  87%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  88%  |                                                                              |=============================================================         |  87%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  88%  |                                                                              |=================================================================     |  93%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |===============================================================       |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |==============================================================        |  88%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |=============================================================         |  87%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |==================================                                    |  49%  |                                                                              |=========================================                             |  59%  |                                                                              |========================================                              |  58%  |                                                                              |=========================================                             |  58%  |                                                                              |========================================                              |  58%  |                                                                              |==================================================                    |  71%

# compare to the previous estimate
summary(BOOT)
#> $name
#> [1] "OUf"
#> 
#> $DOF
#>      mean      area diffusion     speed 
#>  3.721511  2.942958  6.655325 12.563648 
#> 
#> $CI
#>                                           low          est         high
#> area (square kilometers)          4917.699762 24355.095361 59038.545100
#> τ (days)                             2.574366     4.433421     7.634974
#> speed (kilometers/day)               8.403384    11.589090    14.769168
#> diffusion (square kilometers/day)   83.886562   214.735347   406.007849
#> 
# }