Specify, fit, and select continuous-time movement models

These functions allow one to propose hypothethical movement models (with initial estimates), fit those models to the data, and select among those models via an information criteria. The fitting functions wrap around optim and ctmm.loglike to maximize the likelihood function of continuous-time movement models described in Fleming et al (2014) and Fleming et al (2015), given 2D animal tracking data.

ctmm(tau=NULL,isotropic=FALSE,range=TRUE,circle=FALSE,error=FALSE,axes=c("x","y"),...)

ctmm.loglike(data,CTMM,REML=FALSE,verbose=FALSE)

ctmm.fit(data,CTMM=ctmm(),method="ML",control=list(maxit=.Machine$integer.max),
trace=FALSE,...)

ctmm.select(data,CTMM,verbose=FALSE,level=0.99,IC="AICc",trace=FALSE,...)

Arguments

tau

Array of autocorrelation timescales explained below.

isotropic

A Boolean denoting whether or not the animal's covariance is circular or elliptical.

range

A Boolean denoting whether or not the movement model has a finite range.

circle

The period it takes the animal to stochastically circle its mean location.

error

A Boolean denoting whether or not to use annotated telemetry error estimates or an estimate of the error's standard deviation if the data are not annotated with error estimates or when \(HDOP=1\).

axes

Spatial dimensions of the movement model.

data

Timeseries data represented as a telemetry object.

CTMM

A ctmm movement-model object containing the initial parameter guesses conforming to the basic structure of the model hypothesis. ctmm.select can accept a list of such objects.

REML

Use residual maximum likelihood if TRUE. Not recommended.

verbose

Return additional information. See "Value" below.

method

Use residual maximum likelihood if "REML". Not recommended.

control

An optional argument to be passed to optim, but with parscale overwritten with reasonable defaults.

trace

Report progress updates.

...

Further arguments passed to ctmm.fit or optim.

level

Attempt to simplify a model if a feature's non-existence falls within this level of confidence interval.

IC

Information criteria used. Only "AICc" is currently supported.

Details

Model fitting and selection first requires a prototype model with guesstimated parameters (i.e., Brownian motion with a particular diffusion rate). The initial ctmm parameter guess can be generated by the output of ctmm.guess, variogram.fit or manually specified with the function ctmm(...), where the argument tau is explained below and additonal model options described in vignette("ctmm").

By default, tau is an ordered array of autocorrelation timescales. If length(tau)==0, then an IID bi-variate Gaussian model is fit to the data. If length(tau)==1, then an Ornstein-Uhlenbeck (OU) model (Brownian motion restricted to a finite home range) is fit the data, where tau is the position autocorrelation timescale. tau=Inf then yields Brownian motion (BM). If length(tau)==2, then the OUF model (continuous-velocity motion restricted to a finite home range) is fit to the data, where tau[1] is again the position autocorrelation timescale and tau[2] is the velocity autocorrelation timescale. tau[1]=Inf then yields integrated Ornstein-Uhlenbeck (IOU) motion, which is a spatially unrestricted continuous-velocity process.

Model selection in ctmm.select proceeds by first fitting the initial model guess, and then attempting to simplify the autocorrelation model and complexify the deterministic (mean) model until the information criteria fails to improve. The intent of working in these directions is to avoid fitting trends to autocorrelation. Note that simpler models in a nested hierarchy will only be attempted if they appear credible, which can be adjusted with the level argument. level=1 will, therefore, always attempt a simpler model.

Value

The function ctmm returns a prototype ctmm movement-model object. By default, ctmm.loglike returns the log-likelihood of the model CTMM. ctmm.fit (and ctmm.loglike with verbose=TRUE) returns the maximum likelihood ctmm movement-model object with all of the components of CTMM plus the components listed below. ctmm.select returns the best model by default or the list of attempted models if verbose=TRUE.

AICc

The corrected Akaike information criterion, under the assumption of a linear model. Not all model parameters here are linear, but the linear AICc is still implemented rather than AIC because AICc is known to behave well for analogous discrete-time models. Also, following the arguments of Burnham & Anderson (2002), AICc should be an improvement over AIC here because the model parameters are not extremely nonlinear, etc..

loglike

The log-likelihood.

sigma

The maximum likelihood variance/covariance estimate (possibly debiased). For the endlessly diffusing BM and IOU processes, this is instead the diffusion rate estimate.

mu

The maximum likelihood stationary mean vector estimate.

COV.mu

The covariance matrix of the vector mu, assuming that the point estimate sigma is good.

DOF.mu

The effective number of degrees of freedom in the estimate of mu, assuming that the point estimate of tau is good. This can be much smaller than length(data$t) if the data are autocorrelated.

COV

Covariance of the parameter vector c(sigma,tau,circle), as derived from hessian, and where sigma is parameterized in terms of its standard area, eccentricity, and angle of orientation. Typically, sigma's area parameter is extremely correlated to tau[1], and sequential components of tau are slightly correlated.

References

C. H. Fleming, J. M. Calabrese, T. Mueller, K.A. Olson, P. Leimgruber, and W. F. Fagan. (2014). From fine-scale foraging to home ranges: A semi-variance approach to identifying movement modes across spatiotemporal scales. http://www.jstor.org/discover/10.1086/675504

C. H. Fleming and Y. Subasi and J. M. Calabrese. (2015). A maximum-entropy description of animal movement. http://journals.aps.org/pre/abstract/10.1103/PhysRevE.91.032107

Note

Fitting without telemetry errors involves fewer parameters and is much faster than fitting with telemetry errors, as the variance/covariance can be profiled analytically without numerical optimization. Therefore, fitting without telemetry errors, if possible, can provide improved initial estimates of tau and sigma for later fitting with telemetry errors.

The default optim method here is "Nelder-Mead" with the largest allowable number of maximum iterations. Annecdotally, the faster "BFGS" and "L-BFGS-B" methods often perform poorly on these types of problems.

The AICs of endlessly diffusing models like BM and IOU cannot be easily compared to the AICs of range resident models like bivariate Gaussian, OU, and OUF. Endlessly diffusing models have to be conditioned off of an initial state, which we derive in ctmm by taking the large range limit of a range-restricted process. I.e., BM is the limit OU(Inf) and IOU(tau) is the limit OUF(Inf,tau). Using comparable likelihood functions would give up statistical efficiency and the objective prior.

Prior to v0.3.2, ctmm.select would consider every possible model. This is no longer feasible with current versions of ctmm, as the number of possible models has grown too large.

See also

ctmm.guess, optim, summary.ctmm, variogram.fit.

Examples