Draw a random model-fit from the sampling distribution

This function generates random model-fit statistics from the sampling distribution of a given ctmm movement model and sampling schedule. If fast=FALSE, the results are exact, though slow to evaluate. Else if fast=TRUE, the central-limit theorem is invoked.

emulate(object,...)

# S3 method for ctmm
emulate(object,data=NULL,fast=FALSE,...)

# S3 method for telemetry
emulate(object,CTMM,fast=FALSE,...)

Arguments

object
telemetry data or ctmm model object.

CTMM
A ctmm movement-model object.

data
Optional telemetry object for exact results.

fast
Whether or not to invoke the central-limit theorem.

...
Arguments passed to ctmm.fit.

Details

Given fast=FALSE, which requires the data argument specified, new data are simulated from the CTMM movement model with the same sampling schedule and error structure as data. A new model, of the same structure as CTMM, is then fit to the simulated data and returned.

Given fast=TRUE, a model-fit object is sampled from the central-limit distribution, using the covariance estimates within CTMM. Strictly positive parametes, such as area, are log-transformed prior to the normal approximation. Note that this faster method does not adjust for bias.

Value

A ctmm movement model with the same structure as CTMM.

See also