Export ctmm data formats

Functions to export ctmm data formats into common sp, raster, and ESRI formats.

# S4 method for UD
raster(x,DF="CDF",...)

SpatialPoints.telemetry(object,...)

SpatialPolygonsDataFrame.UD(object,level.UD=0.95,level=0.95,...)

writeShapefile(object,...)

# S3 method for UD
writeShapefile(object,folder,file=UD@info$identity,level.UD=0.95,level=0.95,...)

Arguments

x

UD object.

DF

Rasterize the probability density function "PDF", probability mass function "PMF", or cumulative distribution function "CDF".

object

telemetry or UD object.

level.UD

Confidence level of the UD area. I.e., the 50% core home range would be given by level.UD=0.50.

level

Confidence level for the magnitude of the above area. I.e., the 95% CI of the core home range area.

folder

Character name of folder for shapefile.

file

Character name of files for shapefile.

...

Optional arguments passed to writeOGR.

Details

writeShapefile.UD writes a shapefile to disk, with polygons corresponding to the low, ML, and high home-range area estimates.

Value

raster.UD returns a raster of the maximum likelihood (ML) distribution function DF. DF=PDF gives the average probability density per cell, DF=PMF gives the total probability per cell, and DF=CDF gives the cumulative probability.

SpatialPoints.telemetry returns a spatialPoints object for the x-y locations.

SpatialPolygonsDataFrame.UD returns a SpatialPolygonsDataFrame of the low, ML, and high home-range area estimates, in the appropriate order for plotting.

See also

akde, as.telemetry, occurrence.