Extent

Functions to calculate the \((x,y)\) plotting extent (or bounding box) of various ctmm objects or list of such objects, for use when plotting multiple ctmm objects.

# S4 method for telemetry
extent(x,level=1,...)

# S4 method for ctmm
extent(x,level=0.95,level.UD=0.95,...)

# S4 method for UD
extent(x,level=0.95,level.UD=0.95,...)

# S4 method for variogram
extent(x,level=0.95,threshold=2,...)

# S4 method for list
extent(x,...)

Arguments

x
A telemetry, ctmm, or UD object.

level
For telemetry objects, this is the fraction of locations bounded, according to two-sided quantiles. For ctmm and UD objects, this is confidence level for the magnitude of the utilization area circumscribed by level.UD.

level.UD
Confidence level of the UD area. I.e., the 50% core home range would be given by level.UD=0.50.

threshold
Limit ylim to threshold times the maximum semi-variance, even if the level confidence intervals exceed this amount.

...
Optional arguments for future extensions.

Details

Returns a data.frame with columns x and y with rows min and max. See vignette('akde') for an example of extent used to plot multiple UDs on the same scale.

See also