Calculate a range distribution estimate

Estimates the range distributions and suitability from telemetry data and a continuous-time movement model.

homerange(data,CTMM,method="AKDE",...)

agde(data=NULL,CTMM=NULL,R=list(),variable="utilization",error=0.001,res=100,grid=NULL,
     ...)

suitability(data=NULL,CTMM=NULL,R=list(),level=0.95,grid=NULL,log=FALSE,...)

Arguments

data

2D timeseries telemetry data represented as a telemetry object.

CTMM

A ctmm movement model from the output of ctmm.fit.

method

Which range distribution method to use. Can be "AKDE" or "AGDE".

...

Arguments passed to the method call or bandwidth.

R

A named list of raster covariates if CTMM contains an RSF model

variable

Not yet supported.

error

Target probability error.

res

Number of grid points along each axis, relative to the location covariance.

grid

Grid specification via raster, UD, or list of arguments (See akde for details).

level

Confidence level for output confidence intervals.

log

Calculate the log(suitability).

Details

homerange is a wrapper function that calls either akde or agde. Please consult akde for further details on method="AKDE".

suitability calculates a suitability raster from an rsf.fit object. Population RSF fit objects calculated from mean will produce a suitability estimate of the population.

agde calculates autocorrelated Gaussian and RSF home-range areas.

Value

homerange and agde return a UD object. suitability returns a brick object.

Author

C. H. Fleming.

See also