Compare empirical and theoretical intensity (resource-selection) functions [IN DEVELOPMENT]

This function plots the empirical and theoretical intensity functions with respect to a covariate of interest.

intensity(data,UD,RSF,R=list(),variable=NULL,empirical=FALSE,level=0.95,ticks=TRUE,
          smooth=TRUE,interpolate=TRUE,...)

Arguments

data

A telemetry object.

UD

A UD object generated by akde from the same telemetry object as data. If weights were optimized in akde, then they will be adopted by intensity.

RSF

An iRSF model-fit object from rsf.fit or rsf.select.

R

A named list of rasters or time-varying raster stacks [NOT TESTED] to fit Poisson regression coefficients to (under a log link).

variable

Variable of interest from names(R).

empirical

Plot an empirical estimate of \(\log\lambda\) [IN DEVELOPMENT].

level

Confidence level for intensity function estimates.

ticks

Demark used resource values atop the plot.

smooth

Apply location-error smoothing to the tracking data before regression.

interpolate

Whether or not to interpolate raster values during extraction.

...

Arguments passed to plot.

Details

With resepct to the Poisson point process likelihood \(L(\lambda)=\frac{\lambda(x,y)}{\iint \lambda(x',y') \, dx' dy'}\), the formula object of a ctmm iRSF model corresponds to the covariate dependence of \(\log(\lambda)\), which is typically of the form \(\boldsymbol{\beta} \cdot \mathbf{R}\). intensity plots both empirical (black) and theoretical (red) estimates of the log-intensity (or log-selection) function \(\log(\lambda)\) as a function of the covariate variable, which provides a visualization of what the true formula looks like and how the fitted model compares. The empirical estimate is semi-parametric, in that it assumes that RSF is correct for all variables other than variable.

Note

Only relative differences in \(\log(\lambda)\) are meaningful.

See also