Log transformation of parameter estimates and their uncertainties

Methods for log transforming individual parameter estimates and their uncertainty estimates for use in meta-analytic regression, and then back-transforming mean-log parameter estimates back to mean parameter estimates.

Log(x,variable="area",debias=TRUE,...)

Exp(est,VAR.est=0,VAR=0,VAR.VAR=0,variable="area",debias=TRUE,level=0.95,units=TRUE,...)

Arguments

x

A list of UD objects, UD summary objects, or speed objects.

variable

Can be "area", "diffusion", "speed", "tau position", or "tau velocity".

debias

Apply \(\log\chi^2\) and \(\log\chi\) bias corrections if TRUE.

...

Further arguments passed.

est

Point estimate of the mean log-parameter.

VAR.est

Uncertainty in the mean log-parameter estimate (square standard error).

VAR

Variance in the log-parameters.

VAR.VAR

Uncertainty in the log-paramter variance estimate (square standard error).

level

Confidence level for parameter estimates.

units

Convert result to natural units.

Value

Log returns a list with two slots, log and VAR.log, corresponding to the point estimates and variance estimates of the logged variables.

Exp returns a confidence intervals for the back-transformed mean parameter estimate.

Author

C. H. Fleming.

See also

Examples

# \donttest{
# load package and data
library(ctmm)
data(buffalo)

# fit movement models
FITS <- AKDES <- list()
for(i in 1:length(buffalo))
{
  GUESS <- ctmm.guess(buffalo[[i]],interactive=FALSE)
  # use ctmm.select unless you are certain that the selected model is OUF
  FITS[[i]] <- ctmm.fit(buffalo[[i]],GUESS)
}

# calculate AKDES on a consistent grid
AKDES <- akde(buffalo,FITS)
#> Default grid size of 3 minutes chosen for bandwidth(...,fast=TRUE).
#> Default grid size of 2 minutes chosen for bandwidth(...,fast=TRUE).
#> Default grid size of 2 minutes chosen for bandwidth(...,fast=TRUE).
#> Default grid size of 3 minutes chosen for bandwidth(...,fast=TRUE).
#> Default grid size of 2 minutes chosen for bandwidth(...,fast=TRUE).
#> Default grid size of 5 minutes chosen for bandwidth(...,fast=TRUE).

# extract 95% areas
AREAS <- lapply(AKDES,summary)

# log transform for further meta-analysis
LOG <- Log(AREAS)

LOG
#>             log    VAR.log
#> Cilla  19.77354 0.05668715
#> Gabs   20.15899 0.26192932
#> Mvubu  19.74900 0.07487245
#> Pepper 20.47721 0.06594917
#> Queen  19.90336 0.11312026
#> Toni   19.65385 0.05784754
# }