Average autocorrelated kernel density estimates

This function calculates a weighted average of autocorrelated kernel density home-range estimates from akde. The point estimates are correct, but the confidence-interval calculation is not yet complete.

# S3 method for UD
mean(x,weights=NULL,...)

Arguments

x
A list of UDs calculated on the same grid.

weights
A vector of numeric weights with the same length as x, specifying the relative frequency of each distribution in x.

...
Additional arguments for future use.

Details

By default, uniform weights are used (weights=rep(1,length(x))). This can be sensible for averaging over individuals. For averaging over periods of time, users should consider weighting by the proportion of time spent in each distribution. For example, if an animal spends 4 months in its winter range, x[[1]], and 7 months in its summer range, x[[2]], then the annual range (sans migration corridor) would be calculated with weights=c(4,7).

All UDs need to be calculated on the same grid (see overlap for an example).

Value

Returns a UD object: a list with the sampled grid line locations r$x and r$y, the extent of each grid cell dr, the probability density and cumulative distribution functions evaluated on the sampled grid locations PDF & CDF, the optimal bandwidth matrix H, and the effective sample size of the data in DOF.H.

Note

The confidence intervals are not yet calculated correctly.

See also