Compute a number-weighted average of variogram objects

This function takes a list of variogram objects and calculates its number-weighted average variogram.

# S3 method for variogram
mean(x,...)

Arguments

x

A variogram object or list of such objects to be averaged.

...

Additional variograms if specified individually.

Value

Returns a variogram object which is a dataframe containing the lag, the semi-variance estimate at that lag, and the approximate degrees of freedom associated with the semi-variance estimate.

References

C. H. Fleming, J. M. Calabrese, T. Mueller, K.A. Olson, P. Leimgruber, and W. F. Fagan. (2014). From fine-scale foraging to home ranges: A semi-variance approach to identifying movement modes across spatiotemporal scales. http://www.jstor.org/discover/10.1086/675504.

Note

Variogram averaging should only be used when there is a degree of similarity across individual variograms.

See also

plot.variogram, variogram.

Examples