Meta-analysis of movement-model parameters

These functions estimate population-level parameters from individual movement models and related estimates, including AKDE home-range areas, while taking into account estimation uncertainty.

meta(x,level=0.95,level.UD=0.95,level.pop=0.95,method="MLE",IC="AICc",boot=FALSE,
     error=0.01,debias=TRUE,units=TRUE,plot=TRUE,sort=FALSE,mean=TRUE,col="black",...)

Arguments

x
A list of ctmm movement-model objects, UD objects, or UD summary output, constituting a sampled population, or a list of such lists, each constituting a sampled sub-population.

level
Confidence level for parameter estimates.

level.UD
Coverage level for home-range estimates. E.g., 50% core home range.

level.pop
Coverage level for the population of individuals.

method
Statistical estimator used---either maximum likelihood estimation ("MLE") or `best linear unbiased estimator' ("BLUE").

IC
Information criterion to determine whether or not population variation can be estimated. Can be "AICc", AIC, or "BIC".

boot
Perform a parametric bootstrap for confidence intervals and first-order bias correction if debias=TRUE.

error
Relative error tolerance for parametric bootstrap.

debias
Apply Bessel's inverse-Gaussian correction if method="MLE", REML if method="BLUE", and an additional first-order correction if boot=TRUE.

units
Convert result to natural units.

plot
Generate a meta-analysis forest plot.

sort
Sort individuals by their point estimates in forest plot.

mean
Include population mean estimate in forest plot.

col
Color(s) for individual labels and error bars.

...
Further arguments passed to plot.

Details

So-far only the meta-analysis of home-range areas is implemented. More details will be provided in an upcomming manuscript.

For both estimator methods, the same underlying model is assumed.

Value

If x constitutes a sampled population, then meta returns a table with rows corresponding to population coverage levels and columns corresponding to confidence levels of the estimates. If x constitutes a list of sampled sub-populations, then meta returns confidence intervals on the sub-population estimate ratios.

Author

C. H. Fleming.

Note

The AICc formula is approximated via the Gaussian relation.

See also