Calculate a Kriged occurrence disribution estimate

This function calculates an occurrence distribution from telemetry data and a continuous-time movement model.

occurrence(data,CTMM,H=0,res.time=10,res.space=10,grid=NULL,
cor.min=0.5,dt.max=NULL)

Arguments

data
2D timeseries telemetry data represented as a telemetry object.

CTMM
A ctmm movement model from the output of ctmm.fit.

H
Optional additional banwidth matrix for future use.

res.time
Number of temporal grid points per median timestep.

res.space
Number of grid points along each axis, relative to the average diffusion (per median timestep) from a stationary point.

grid
Optional grid specification with columns labeled x and y. Not yet supported.

cor.min
Minimum gap correlation for Kriging interpolation (alternative to dt.max).

dt.max
Maximum gap size for Kriging interpolation (alternative to cor.min).

Value

Returns a UD object containing the sampled grid line locations x and y, the probability density and cumulative distribution functions evaluated on the sampled grid locations PDF & CDF, the optional bandwidth matrix H, and the area of each grid cell dA.

References

C. H. Fleming and W. F. Fagan and T. Mueller and K. A. Olson and P. Leimgruber and J. M. Calabrese. (2016). Estimating where and how animals travel: An optimal framework for path reconstruction from autocorrelated tracking data. Ecology, DOI:10.1890/15-1607.

Note

Large gaps have a tendency to blow up the estimate, and can be avoided with the cor.min or dt.max arguments.

In the case of coarse grids, the value of PDF in a grid cell actually corresponds to the average probability density over the entire rectangular cell.

See also

akde, raster,UD-method

Examples