Calculate the overlap between two stationary distributions

This function calculates a useful measure of similarity between distributions known as the Bhattacharyya coefficient in statistics and simply the fidelity or overlap in quantum and statistical mechanics. It is roughly speaking the ratio of the intersection area to the average individual area. When applied to ctmm objects, this function returns the overlap of the two Gaussian distributions. When applied to telemetry or (aligned) UD objects with corresponding movement models, this function returns the overlap of their (autocorrelated) kernel density estimates.

overlap(object,CTMM=NULL,level=0.95,debias=FALSE,...)

Arguments

object
A list of ctmm fit, telemetry, or aligned UD objects to compare.

CTMM
A list of ctmm fit objects corresponding to object list.

level
The confidence level desired for the output.

debias
Approximate debiasing of the overlap. (TESTING!)

...
Additional arguments relevant for akde, such as res and weights.

Value

A table of confidence intervals on the overlap estimate. A value of 1 implies that the two distributions are identical, while a value of 0 implies that the two distributions share no area in common. Corresponding ctmm objects are necessary to provide confidence intervals and debiasing for the point esitmates.

Note

Uncertainties in CTMM1 and CTMM2 are propagated into the overlap estimate under the approximation that the Bhattacharyya distance is a chi-square random variable. Debiasing makes further approximations noted in Winner & Noonan et al (2018).

See also

akde, ctmm.fit

Examples

# Load package and data
library(ctmm)
data(buffalo)

# fit models for first two buffalo
GUESS <- lapply(buffalo[1:2], function(b) ctmm.guess(b,interactive=FALSE) )
FITS <- lapply(1:2, function(i) ctmm.fit(buffalo[[i]],GUESS[[i]]) )
names(FITS) <- names(buffalo[1:2])

# Gaussian overlap between these two buffalo
overlap(FITS)

#> , , low
#> 
#>           Cilla      Gabs
#> Cilla 1.0000000 0.8877895
#> Gabs  0.8877895 1.0000000
#> 
#> , , ML
#> 
#>           Cilla      Gabs
#> Cilla 1.0000000 0.9840319
#> Gabs  0.9840319 1.0000000
#> 
#> , , high
#> 
#>       Cilla Gabs
#> Cilla     1    1
#> Gabs      1    1
#> 


# AKDE overlap between these two buffalo
overlap(buffalo[1:2],FITS)

#> , , low
#> 
#>           Cilla      Gabs
#> Cilla 1.0000000 0.8639276
#> Gabs  0.8639276 1.0000000
#> 
#> , , ML
#> 
#>           Cilla      Gabs
#> Cilla 1.0000000 0.9411288
#> Gabs  0.9411288 1.0000000
#> 
#> , , high
#> 
#>           Cilla      Gabs
#> Cilla 1.0000000 0.9876048
#> Gabs  0.9876048 1.0000000
#>