Plotting methods for variogram objects.

Produces simple plots of varigram objects (semi-variance vs. time lag) and model semi-variance functions, with approximate confidence intervals around the semi-variance estimates.

# S3 method for variogram
plot(x,CTMM=NULL,level=0.95,fraction=0.5,col="black",col.CTMM="red",...)

# S4 method for variogram
zoom(x,fraction=0.5,...)

Arguments

x

A variogram object calculated using variogram.

CTMM

A ctmm movement model object in the same format as the output of ctmm.fit or variogram.fit.

level

Confidence level of confidence bands (95% default CIs). Can be an array.

fraction

The proportion of the variogram object, variogram, that will be plotted. By convention, half is shown. The tail end is generally garbage.

col

Color for the empirical variogram. Can be an array.

col.CTMM

Color for the model. Can be an array.

…

Additional plot function parameters.

Value

Returns a plot of semi-variance vs. time lag, with the empirical variogram in blue and the ctmm semi-variance function in red if specified. zoom includes a log-scale zoom slider to manipulate fraction.

References

C. H. Fleming, J. M. Calabrese, T. Mueller, K.A. Olson, P. Leimgruber, and W. F. Fagan. (2014). From fine-scale foraging to home ranges: A semi-variance approach to identifying movement modes across spatiotemporal scales. http://www.jstor.org/discover/10.1086/675504

Note

The errors of the empirical variogram are correlated. Smooth trends are not necessarily significant.

See also

ctmm.fit, plot, variogram, variogram.fit.

Examples

# Load package and data
library(ctmm)
data(buffalo)

# Extract movement data for a single animal
cilla <- buffalo[[1]]

# Calculate variogram
SVF <- variogram(cilla)

#> Error in NextMethod("["): 'NextMethod' called from an anonymous function

# Plot the variogram
plot(SVF)

#> Error in plot(SVF): object 'SVF' not found