Estimate the average speed of a tracked animal

Given a ctmm movement model and telemetry data, speed simulates multiple realizations of the animal's trajectory and estimates the average speed, which is proportional to distance traveled. Both tortuosity (non straight-line motion between the data) and telemetry error can be accounted for. Given only a ctmm movement model and no data, speed calculates the average speed of the Gaussian movement process.

speed(object,...)

# S3 method for ctmm
speed(object,data=NULL,level=0.95,units=TRUE,prior=TRUE,fast=TRUE,cor.min=0.5,dt.max=NULL,
      error=0.01,cores=1,...)

# S3 method for telemetry
speed(object,CTMM,level=0.95,units=TRUE,prior=TRUE,fast=TRUE,cor.min=0.5,dt.max=NULL,
      error=0.01,cores=1,...)

Arguments

object
A ctmm movement-model or telemetry object, which requires an additional CTMM argument.

data
Optional telemetry object on which the simulations will be conditioned.

CTMM
Movement model object.

level
Confidence level to report on the estimated average speed.

units
Convert result to natural units.

prior
Account for model uncertainty.

fast
Whether or not to invoke the central-limit theorem (see emulate).

cor.min
Velocity correlation threshold for skipping gaps.

dt.max
Absolute gap sizes to skip (in seconds), alternative to cor.min.

error
Target (relative) standard error.

cores
Number of simulations to run in parallel. cores=NULL will use all cores, while cores<0 will reserve abs(cores).

...
Arguments passed to emulate.

Details

The cor.min or dt.max arguments are used to constrain the estimate to be derived from simulations near the data, and therefore ensure that the estimate is more reflective of the data than the model.

Value

Returns the estimated mean speed of the sampled trajectory with CIs by default. If level=NULL, then the ensemble of mean speeds is returned instead.

Note

The average speeds estimated here are mean speeds. The speeds reported by summary.ctmm are root-mean-square (RMS) speeds. These quantities are sometimes proportional, but not equivalent.

See also