Summarize a range distribution

This function returns a list of biologically interesting parameters in human readable format, as derived from an autocorrelated kernel density estimate.

# S3 method for UD
summary(object,level.UD=0.95,level=0.95,units=TRUE,...)

Arguments

object
An akde autocorrelated kernel-density estimate from the output of akde.

level.UD
Confidence level for the home-range area. E.g., the 50% core area.

level
Confidence level for the above area estimate. E.g., the 95% confidence interval of the 50% core area.

units
Convert result to natural units.

...
Unused options.

Value

A matrix with low, maximum likelihood, and high estimates for the following parameters:

area

The home-range area with fraction of inclusion level.UD. E.g., the 50% core home range is estimated with level.UD=0.50, and 95% confidence intervals are placed on that area estimate with level=0.95.

References

C. H. Fleming, J. M. Calabrese. A new kernel-density estimator for accurate home-range and species-range area estimation. Methods in Ecology and Evolution, 8:5, 571-579 (2016).

Note

Prior to ctmm v0.3.1, AKDEs included only errors due to autocorrelation uncertainty, which are insignificant in cases such as IID data. Starting in v0.3.1, akde calculated an effective sample size DOF.H and used this to estimate area uncertainty under a chi-square approxmation. Starting in v0.3.2, this method was improved to use DOF.area in the Gaussian reference function approximation.

See also

Examples

# Load package and data
library(ctmm)
data(buffalo)

# Extract movement data for a single animal
Cilla <- buffalo$Cilla

# Fit a movement model
GUESS <- ctmm.guess(Cilla,interactive=FALSE)
FIT <- ctmm.fit(Cilla,GUESS)

# Estimate and summarize the AKDE
UD <- akde(Cilla,FIT)
summary(UD)

#> $DOF
#>      area bandwidth 
#>  22.52394  32.53579 
#> 
#> $CI
#>                               low       ML     high
#> area (square kilometers) 227.5339 360.8605 524.4376
#>