Estimate UERE from calibration data

A function for estimating the User Equivalent Range Error (UERE) of a GPS device from calibration data.

uere(data,diagnostic=FALSE)

Arguments

data

Telemetry object or list of telemetry objects corresponding to calibration data with Movebank column GPS.HDOP.

diagnostic

Return extra details, such as the UERE's 95% confidence interval and a plot.

Details

Often times GPS animal tracking devices return HDOP values but do not specifiy the device's UERE necessary to transform the HDOP values into absolute errors. This function allows users to estimate the UERE from calibration data, where the device was left fixed over a period of time. Otherwise, when error=TRUE in ctmm, ctmm.fit will estimate the UERE simultaneously with the movement model, which is less reliable than using calibration data.

Value

The UERE estimate.

Note

The GPS device should be fixed during calibraiton.