Calculate an empirical variogram from movement data

This function calculates the isotropic empirical variogram of two-dimensional tracking data for visualizing stationary (time-averaged) autocorrelation structure. One of two algorithms is used. The slow \(O(n^2)\) algorithm is based upon the lag-weighted variogram of Fleming et al (2014). Additional modifications have also been included to accommodate any drift in the sampling rate. The fast \(O(n \log n)\) algorithm is based upon the FFT method of Marcotte (1996), with some tweaks to better handle irregularly sampled data. Both methods reduce to the standard “method of moments” estimator in the case of evenly scheduled data, even with missing observations, but they produce slightly different outputs for irregularly sampled data.

variogram(data,dt=NULL,fast=TRUE,CI="Markov",axes=c("x","y"))

Arguments

data

telemetry data object of the 2D timeseries data.

dt

Lag bin width. An ordered array will yield a progressive coarsening of the lags. Defaults to the median sampling interval.

fast

Use the lag-weighted algorithm if FALSE or the FFT algorithm if TRUE. The slow algorithm outputs a progress bar.

CI

Defaults to only consider non-overlapping lags independent. All unique lags are considered independent with "IID".

axes

Array of axes to calculate an average (isotropic) variogram for.

Details

If no dt is specified, the median sampling interval is used. This is typically a good assumption for most data, even when there are gaps. A dt coarser than the sampling interval may bias the variogram (particuarly if fast=TRUE) and so this should be reserved for poor data quality.

For irregularly sampled data, it may be useful to provide an array of time-lag bin widths to progressively coarsen the variogram. I.e., if you made the very bad choice of changing your sampling interval on the fly from dt1 to dt2, where dt1 \(<\) dt2, the an appropriate choice would be dt=c(dt1,dt2). On the other hand, if your sampling is itself a noisy process, then you might want to introduce larger and larger dt components as the visual appearance of the variogram breaks down with increasing lags. Alternatively, you might try the fast=FALSE option or aggregating multiple individuals with mean.variogram.

In standard variogram regression treatments, all lags are considered as independent (CI="IID") for the purposes of confidence-interval estimation, even if they overlap in time. However, in high resolution datasets this will produce vastly underestimated confidence intervals. Therefore, the default CI="Markov" behavior is to consider only the maximum number of non-overlapping lags in calculating confidence intervals.

Value

Returns a variogram object (class variogram) which is a dataframe containing the time-lag, lag, the semi-variance estimate at that lag, SVF, and the approximate number of degrees of freedom associated with that semi-variance, DOF, with which its confidence intervals can be estimated.

References

C. H. Fleming, J. M. Calabrese, T. Mueller, K.A. Olson, P. Leimgruber, and W. F. Fagan. (2014). From fine-scale foraging to home ranges: A semi-variance approach to identifying movement modes across spatiotemporal scales. http://www.jstor.org/discover/10.1086/675504.

D. Marcotte. (1996). Fast variogram computation with FFT. Computers and Geosciences 22(10), 1175-1186.

See also

vignette("variogram"), mean.variogram, plot.variogram, variogram.fit.

Examples

#Load package and data
library(ctmm)
data(buffalo)

#Extract movement data for a single animal
cilla <- buffalo[[1]]

#Calculate variogram
SVF <- variogram(cilla)

#> Error in NextMethod("["): 'NextMethod' called from an anonymous function

#Plot the variogram
plot(SVF)

#> Error in plot(SVF): object 'SVF' not found