Import, coerce, and summarize MoveBank data

Functions to import MoveBank csv files, data.frame, and Move objects, coerce them into telemetry objects, and summarize them.

as.telemetry(object,timeformat="",timezone="UTC",projection=NULL,timeout=Inf,na.rm="row",
             mark.rm=FALSE,drop=TRUE,...)

# S3 method for character
as.telemetry(object,timeformat="",timezone="UTC",projection=NULL,timeout=Inf,na.rm="row",
             mark.rm=FALSE,drop=TRUE,...)

# S3 method for data.frame
as.telemetry(object,timeformat="",timezone="UTC",projection=NULL,timeout=Inf,na.rm="row",
             mark.rm=FALSE,drop=TRUE,...)

# S3 method for Move
as.telemetry(object,timeformat="",timezone="UTC",projection=NULL,timeout=Inf,na.rm="row",
             mark.rm=FALSE,drop=TRUE,...)

# S3 method for telemetry
summary(object,...)

Arguments

object
A MoveBank CSV filename, MoveBank data.frame object, or Move object to coarce, or a telemetry object to summarize.

timeformat
Format argument for strptime.

timezone
Timezone argument for strptime.

projection
Optional PROJ.4 projection argument to be fed to project.

timeout
GPS location fix timeout value (seconds). Potentially useful for categorizing accuracy. timeout=max will use the maximum value.

na.rm
If some values are NA in the data frame, are the rows (times) deleted or are the columns (data types) deleted.

mark.rm
Delete Movebank manually marked outliers. Also see outlie.

drop
Only return a telemetry object for one individual if TRUE. Always return a list of telemetry objects if FALSE.

...
Optional arguments to be fed to fread or read.csv, in the case of compressed files.

Details

If no projection argument is specified, a two-point equidistant projection is calculated that should be good for most range resident and migratory species. Global migrations that are not along one geodesic (locally straight line) will probably suffer distortion.

as.telemetry assumes Movebank naming conventions. Sufficient MoveBank columns include individual.local.identifier (or tag.local.identifier), timestamp, location.long and location.lat, while the optional Movebank columns include (E-OBS) eobs.horizontal.accuracy.estimate, (Telonics) GPS.Horizontal.Error, GPS.HDOP, (ARGOS) Argos.orientation, Argos.semi.minor and Argos.semi.major or Argos.location.class, etc..

The GPS.HDOP value requires the device's UERE value to be translated into telemetry errors. The UERE represents the RMS error given ideal (\(HDOP=1\)) satellite conditions. Therefore, if your device is purported to have an accuracy of 10 meters, then the UERE is likely around 10 meters. Consult your device manual and manufacturer for a specific value or see uere for UERE calibration.

Value

as.telemetry returns a single telemetry object or list of telemetry objects if multiple animals are identified.

See also