Import, coerce, and summarize MoveBank data

Functions to import MoveBank csv files, data.frame, and Move objects, coerce them into telemetry objects, and summarize them.

as.telemetry(object,timeformat="",timezone="UTC",projection=NULL,UERE=NULL,...)

# S3 method for character
as.telemetry(object,timeformat="",timezone="UTC",projection=NULL,UERE=NULL,...)

# S3 method for data.frame
as.telemetry(object,timeformat="",timezone="UTC",projection=NULL,UERE=NULL,...)

# S3 method for Move
as.telemetry(object,timeformat="",timezone="UTC",projection=NULL,UERE=NULL,...)

# S3 method for telemetry
summary(object,...)

Arguments

object

A MoveBank CSV filename, MoveBank data.frame object, or Move object to coarce, or a telemetry object to summarize.

timeformat

Format argument for strptime.

timezone

Timezone argument for strptime.

projection

Optional PROJ.4 projection argument to be fed to project.

UERE

"User equivalent range errors" needed to translate DOP values into location uncertainty.

...

Optional arguments to be fed to read.csv.

Details

If no projection argument is specified, a two-point equidistant projection is calculated that should be good for most range resident and migratory species. Global migrations that are not along one geodesic (locally straight line) will probably suffer distortion.

as.telemetry assumes http://rstudio.com. Sufficient MoveBank columns include individual.local.identifier (or tag.local.identifier), timestamp, location.long and location.lat, while the optional Movebank columns are eobs.horizontal.accuracy.estimate and GPS.HDOP. In the future, we will be supporting more GPS and ARGOS columns, including speed, heading, further error classifications, etc.

The GPS.HDOP value requires the device's UERE value to be translated into telemetry errors. The UERE represents the RMS error given ideal satellite conditions. Therefore, if your device is purported to have an accuracy of 2-10 meters, then the UERE is likely around 2 meters. Consult your device manual and manufacturer for a specific value or see uere for UERE calibration.

Value

as.telemetry returns a single telemetry object or list of telemetry objects if multiple animals are identified.

See also

plot.telemetry, SpatialPoints.telemetry, uere.