ctmm FAQ

Frequently asked questions for the ctmm package.

Details

General recommendations

Never edit your CSV in Microsoft Excel. The dates will be reformatted incorrectly.

Do not save workspaces between sessions. They can become corrupted.

Upgrade R to the latest version and update all of your packages.

Work through the vignettes vignette("variogram") and vignette("akde").

Check the ctmm website to see if there is a beta release that fixes any bugs.

Check the ctmm user's group.

Help installing packages on Linux

These are the packages I needed in Ubuntu: sudo apt install fftw3 fftw3-dev libgdal1-dev libproj-dev libgeos-dev r-base-core

See Xianghui Dong's installation tips here.

ctmm.select only returns one model with verbose=TRUE The newer version of ctmm.select is much more intelligent, necessitated by the exploding number of new model permutations. Any simpler models were judged to be so unlikely that they were not attempted. You can set level=1 to force alternative model fitting, but the additional models should lose badly in model selection (please contact me if I am wrong).

rdb database corruption, "could not find function", and other weird errors R likely did not install or load the package correctly. E.g., some files may have failed to overwrite previous versions. Uninstall ctmm, restart R without saving the workspace/session, and install ctmm again.

Infinite recursion and stack overflow errors ctmm has no recursive functions, so I am not exactly sure what causes this error, but it only occurs with certain versions of R on certain computer architectures. There are several solutions that have worked for people, including restarting R in a fresh session and updating their software. Alternatively:

Reboot your computer.

Increase the allowed number of nested expressions within R via options(expressions=10000) or some other large number.

Try a different computer.

manipulate::isAvailable is not found You probably have an outdated copy of the manipulate package installed. Update R to the latest version and then update all of your packages. This seems to happen frequently with the MacOS release of R.

manipulate panel does not popup in variogram.fit or zoom Click the gear icon in the upper-left corner of the plot window.

plot complains about the datatype or has weird errors Namespace collision sometimes occurs from raster or sp against move and ctmm. Restart R and only load the ctmm package.

variogram.fit has no save button Maximize the plot window and/or increase your screen resolution.