ctmm FAQ

Frequently asked questions for the ctmm package.

Details

General recommendations

Work through the vignettes vignette("variogram") and vignette("akde"). Also, see the help file for the method of interest, and its example.

Do not save workspaces between sessions. They will become corrupted over time. In RStudio, go to Tools: Global Options: Workspace, uncheck Restore and set Save to Never.

If RStudio is crashing frequently in Windows (or your display driver is crashing), try setting the rendering engine to Software under Tools : Global Options : General : Advanced : Rendering Engine.

Never edit or save your CSV in Microsoft Excel. The dates will be reformatted incorrectly and inconsistently.

If using Windows, make sure to have the suggested version of ``Rtools'' installed. If using MacOS, make sure to have ``Xcode'' installed. If using Ubuntu, make sure to have ``build-essential'' installed. Otherwise, you can sometimes run into problems when trying to update packages.

Upgrade R to the latest version and update all of your packages.

The development build can be installed via remotes::install_github("ctmm-initiative/ctmm").

Sometimes installing from Github can silently fail to overwrite old files, requiring the package to be manually uninstalled, and then re-installed after restarting.

Stable beta releases between the CRAN release are published here on request.

The ctmm user's group is a good place to find and ask for help.

Bug reports and feature requests can be raised at the Github project page.

Help installing packages on Linux

These are the packages I needed in Ubuntu:sudo apt install ffmpeg fftw3 libfftw3-dev libgdal-dev libgeos-dev libgit2-dev libgmp-dev libgsl-dev libmpfr-dev libproj-dev libnode-dev libudunits2-dev r-base-core

as.telemetry reports abnormal sampling intervals and speeds

Make sure that you have the correct timezone and timeformat arguments specified. Also, see outlie.

rdb database corruption, "could not find function", "cannot coerce class", and other weird errors

R might not have installed or loaded the package correctly---e.g., some files may have failed to overwrite previous versions---or the workspace/session might be corrupted. Uninstall ctmm, restart R without saving the workspace/session, and install ctmm again.

Infinite recursion and stack overflow errors

ctmm has no recursive functions, so I am not exactly sure what causes this error, but it only occurs with certain versions of R on certain computer architectures. There are several solutions that have worked for people, including restarting R in a fresh session and updating their software. Alternatively:

Reboot your computer.

Increase the allowed number of nested expressions within R via options(expressions=10000) or some other large number.

Try a different computer.

plot complains about the datatype or has weird errors

Namespace collision sometimes occurs between raster, sp, move, and ctmm. Either restart R and only load the ctmm package, or run ctmm::plot instead of plot.

North is no longer up after importing data

The default projection in ctmm does not preserve the direction of North, but better preserves distances for elongated distributions. See the projection argument in as.telemetry and the example in projection. The compass function is also useful for pointing north.

projection complains about the datatype and fails

Namespace collision can occur between raster and ctmm. Either restart R and only load the ctmm package, or run ctmm::projection instead of projection.

ctmm.guess has no save button

Maximize the plot window and/or increase your screen resolution.

manipulate panel does not popup in ctmm.guess or zoom

Click the gear icon in the upper-left corner of the plot window.

Gear icon missing in ctmm.guess or zoom

Recent versions of manipulate and/or RStudio seem to have some issues. Sometimes the gear icon does not render unless you re-run the function 2-5 times.

manipulate::isAvailable is not found

You probably have an outdated copy of the manipulate package installed. Update R to the latest version and then update all of your packages. This seems to happen frequently with the MacOS release of R.

Author

C. H. Fleming