Continuous-time movement modeling

Description: Functions for identifying, fitting, and applying continuous-space, continuous-time stochastic movement models to animal tracking data. The package is described in Calabrese & Fleming (2016) and its methods are based on those introduced in Fleming et al (2014-2016) and Péron et al (2016).

Details

Package:

ctmm

Type:

Package

Version:

0.3.6

Date:

2017-03-09

License:

GPL-3

ctmm-FAQ

http://biology.umd.edu/movement.html

http://www2.physics.umd.edu/~hfleming/

https://github.com/ctmm-initiative/ctmm

https://ctmm-initiative.github.io/ctmm/

https://CRAN.R-project.org/package=ctmm

https://groups.google.com/forum/#!forum/ctmm-user

References

J. M. Calabrese, C. H. Fleming. (2016). ctmm: an R package for analyzing animal relocation data as a continuous-time stochastic process. http://onlinelibrary.wiley.com/wol1/doi/10.1111/2041-210X.12559/abstract.

C. H. Fleming, J. M. Calabrese, T. Mueller, K.A. Olson, P. Leimgruber, and W. F. Fagan. (2014). From fine-scale foraging to home ranges: A semi-variance approach to identifying movement modes across spatiotemporal scales. http://www.jstor.org/discover/10.1086/675504.

C. H. Fleming and J. M. Calabrese and T. Mueller and K. A. Olson and P. Leimgruber, and W. F. Fagan (2014). Non-Markovian maximum likelihood estimation of autocorrelated movement processes http://onlinelibrary.wiley.com/doi/10.1111/2041-210X.12176/abstract.

C. H. Fleming and Y. Subasi and J. M. Calabrese. (2015). A maximum-entropy description of animal movement. http://journals.aps.org/pre/abstract/10.1103/PhysRevE.91.032107.

C. H. Fleming and W. F. Fagan and T. Mueller and K. A. Olson and P. Leimgruber, and J. M. Calabrese (2015). Rigorous home-range estimation with movement data: A new autocorrelated kernel-density estimator. http://www.esajournals.org/doi/abs/10.1890/14-2010.1.

C. H. Fleming, W. F. Fagan, T. Mueller, K. A. Olson, P. Leimgruber, and J. M. Calabrese. (2016). Estimating where and how animals travel: An optimal framework for path reconstruction from autocorrelated tracking data. http://onlinelibrary.wiley.com/doi/10.1890/15-1607/full.

G. Péron, C. H. Fleming, R. C. de Paula, J. M. Calabrese. (2016). Uncovering periodic patterns of space use in animal tracking data with periodograms, including a new algorithm for the Lomb-Scargle periodogram and improved randomization tests. https://movementecologyjournal.biomedcentral.com/articles/10.1186/s40462-016-0084-7.

C. H. Fleming, J. M. Calabrese, A new kernel-density estimator for accurate home-range and species-range area estimation, http://onlinelibrary.wiley.com/wol1/doi/10.1111/2041-210X.12673/abstract.