Parametric bootstrap continuous-time movement models

This function allows the point estimates and confidence intervals of an initial estimated movement model to be improved by parametric boostrap, as described in Fleming et al (2018).

ctmm.boot(data,CTMM,method=c('HREML','pREML','pHREML'),error=0.01,cores=1,trace=TRUE,...)

Arguments

data
Timeseries data represented as a telemetry object.

CTMM
A ctmm movement-model object from the output of ctmm.fit containing the initial parameter estimates.

method
Fitting method(s) to use: "ML", "HREML", "pREML", "pHREML", or "REML". See "Description" below.

error
Relative standard error target for bootstrap ensemble estimates and nonlinear iterations.

cores
Number of simulations to run in parallel. cores=NULL will use all cores, while cores<0 will reserve abs(cores).

trace
Report progress updates. Can be among 0:2 with increasing detail.

...
Further arguments passed to ctmm.fit.

Details

ctmm.boot can leverage multiple estimators via the method argument (see ctmm.fit) and as described in Fleming et al (2018), though generally this is only useful if the specified estimators deviate substantially from each other realtive to the prescribed error.

Value

A model fit object with relatively unbiased estimates of home-range area, diffusion rate, and RMS speed (and more accurate CIs than ctmm.fit).

See also