Export ctmm data formats

Functions to export ctmm data formats into common sp, raster, and ESRI formats.

# S4 method for UD
raster(x,DF="CDF",...)

SpatialPoints.telemetry(object,...)

SpatialPointsDataFrame.telemetry(object,...)

SpatialPolygonsDataFrame.telemetry(object,level.UD=0.95,...)

SpatialPolygonsDataFrame.UD(object,level.UD=0.95,level=0.95,...)

# S4 method for UD,character
writeRaster(x,filename,format,DF="CDF",...)

writeShapefile(object,folder,file=NULL,...)

# S3 method for telemetry
writeShapefile(object,folder,file=NULL,error=TRUE,level.UD=0.95,...)

# S3 method for UD
writeShapefile(object,folder,file=NULL,level.UD=0.95,level=0.95,...)

Arguments

x
UD object.

DF
Rasterize the probability density function "PDF", probability mass function "PMF", or cumulative distribution function "CDF".

object
telemetry or UD object.

level.UD
Coverage level of the UD area. I.e., the 50% core home range would be given by level.UD=0.50.

level
Confidence level for the magnitude of the above area. I.e., the 95% CI of the core home range area.

filename
Character name of file for raster file.

format
Character format, if not inferred from filename extension (see writeRaster.)

folder
Character name of folder for shapefile.

file
Character name of files for shapefile.

error
Export error circles/ellipses as polygons if TRUE.

...
Optional arguments passed to writeRaster, writeOGR, etc..

Details

writeRaster writes a raster file to disk, with pixel values corresponding to the density function.

writeShapefile writes a shapefile to disk, with UD polygons corresponding to the low-CI, point-estimate, and high-CI home-range area estimates.

Value

raster returns a raster of the point-estimate distribution function DF, given a UD object. DF="PDF" gives the average probability density per cell, DF="PMF" gives the total probability per cell, and DF="CDF" gives the cumulative probability.

SpatialPoints.telemetry returns a single spatialPoints object for the x-y locations, with individual identity information lost.

SpatialPointsDataFrame.telemetry returns a SpatialPointsDataFrame with the individual identities recorded in the data frame.

SpatialPolygonsDataFrame.telemetry returns a SpatialPolygonsDataFrame that encodes the location estimate's error circles/ellipses.

SpatialPolygonsDataFrame.UD returns a SpatialPolygonsDataFrame of the low-CI, point-estimate, and high-CI home-range area estimates, in the appropriate order for plotting.

Author

C. H. Fleming and K. Safi.

See also