Average movement models and autocorrelated kernel density estimates

These functions calculate population averages of continuous-time movement models and utilization distributions.

# S3 method for ctmm
mean(x,weights=NULL,sample=TRUE,debias=TRUE,IC="AIC",trace=TRUE,...)

# S3 method for UD
mean(x,weights=NULL,sample=TRUE,...)

Arguments

x

A list of ctmm objects calculated in the same projection or UD objects calculated on the compatible grids.

weights

A vector of numeric weights with the same length as x, specifying the relative frequency of each distribution in x.

sample

x represents a sample of a larger population if TRUE, or the entire statistical population if FALSE.

debias

Include \(\log-\chi^2\) and REML bias corrections.

IC

Model selection criterion for the anisotropy of the distribution of mean locations and covariance matrices.

trace

Report location and autocovariance model selection results.

...

Additional arguments for future use.

Details

When applied to a list of ctmm objects, mean calculates an average movement model with populaton variability estimates. The population model is taken to be multivariate normal and log-normal. The population mean location represents an arithmetic mean, while the population mean home-range areas, RMS speeds, and diffusion rates represent geometric means. Location-error estimates are not correctly averaged yet.

When applied to a list of UD objects, mean calculates a weighted average of autocorrelated kernel density home-range estimates from akde. The point estimates are correct, but the confidence-interval calculation is not yet complete.

By default, uniform weights are used (weights=rep(1,length(x))). This can be sensible for averaging over individuals. For averaging over periods of time, users should consider weighting by the proportion of time spent in each distribution. For example, if an animal spends 4 months in its winter range, x[[1]], and 7 months in its summer range, x[[2]], then the annual range (sans migration corridor) would be calculated with weights=c(4,7).

All UDs need to be calculated on the same grid (see overlap for an example).

Value

When applied to a list of ctmm objects, mean returns a ctmm object with additional population variability parameter estimates.

When applied to a list of UD objects, mean returns a UD object: a list with the sampled grid line locations r$x and r$y, the extent of each grid cell dr, the probability density and cumulative distribution functions evaluated on the sampled grid locations PDF & CDF, the optimal bandwidth matrix H, and the effective sample size of the data in DOF.H.

Author

C. H. Fleming

See also