Spatial selection methods for telemetry objects.

Methods to segment or subset telemety objects based on polygon lasso, rectangular marquee, and time slider selectors.

lasso(object,...)

marquee(object,...)

cleave(object,fraction=0.5,name="CLEFT",...)

Arguments

object

telemetry object or list of such objects.

fraction

Initial split, as fraction of total time period.

name

Name of list to store cleft telemetry objects to.

...

Additional arguments passed to plot.

Details

lasso and marquee allow the user to subset telemetry data into two groups (interior and exterior), based on a hand-drawn polygon lasso or rectangular marquee. cleave allows the user to split the data into two halves at a particular time selected via slider.

Value

lasso and marquee return a named list telemetry objects, twice the length of the input object, where the first half are the interior subsets and the second half are the exterior subsets. cleave stores a similar list of telemetry objects to name on button press.

Author

C. H. Fleming.

See also

Examples

# This example is interactive
if(interactive())
{
  # Load package and data
  library(ctmm)
  data(wolf)

  # Extract wolf Luna
  DATA <- wolf$Luna

  # Select resident data
  SUB <- lasso(DATA)

  # You can now work with the resident and dispersive data separately
  names(SUB)
}