Summarize a continuous-time movement model

This function returns a list of biologically interesting parameters in human readable format, as derived from a continuous-time movement model.

# S3 method for ctmm
summary(object,level=0.95,level.UD=0.95,units=TRUE,IC="AICc",MSPE="position",...)

Arguments

object
A ctmm movement-model object from the output of ctmm.fit.

level
Confidence level for parameter estimates.

level.UD
Coverage level for the Gaussian home-range area.

units
Convert result to natural units.

IC
Information criteria for sorting lists of ctmm objects. Can be "AICc", "AIC", "BIC", "LOOCV", or none (NA). AICc is approximate.

MSPE
Sort models with the same autocovariance structure by the mean square predictive error of "position", "velocity", or not (NA).

...
Unused options.

Value

If summary is called with a single ctmm object output from ctmm.fit, then a list is returned with the effective sample sizes of various parameter estimates (DOF) and a parameter estimate table CI, with low, maximum likelihood, and high estimates for the following possible parameters:

tau

The autocorrelation timescales. tau position is also the home-range crossing timescale.

area

The Gaussian home-range area, where the point estimate has a significance level of level.UD. I.e., the core home range is where the animal is located 50% of the time with level.UD=0.50. This point estimate itself is subject to uncertainty, and is given confidence intervals derived from level.This Gaussian estimate differs from the kernel density estimate of summary.UD. The Gaussian estimate has more statistical efficiency, but is less related to space use for non-Gaussian processes.

speed

The Gaussian root-mean-square (RMS) velocity, which is a convenient measure of average speed but not the conventional measure of average speed (see speed).

If summary is called on a list of ctmm objects output from ctmm.select, then a table is returned with the model names and IC differences for comparison across autocovariance structures. Given non-stationary models and MSPE>0, the mean square prediction error (MSPE) is also returned for comparison across trend structures (with autocovariance structure fixed). For the model names, "IID" denotes the uncorrelated bi-variate Gaussian model, "OU" denotes the continuous-position Ornstein-Uhlenbeck model, and "OUF" denotes the continuous-velocity Ornstein-Uhlenbeck-F model.

Note

Confidence intervals on the autocorrelation timescales assume they are sufficiently greater than zero and less than infinity.

IC="LOOCV" can only be attempted if also specified during ctmm.select, as this argument requires additional calculations.

In ctmm v0.5.1 the MSPE was averaged over all possible times instead of over all sampled times.

In ctmm v0.3.4 the speed estimate was fixed to be the RMS velocity and not \(1/\sqrt{2}\) times the RMS velocity.

See also

Examples

# \donttest{
# Load package and data
library(ctmm)
data(buffalo)

# Extract movement data for a single animal
Cilla <- buffalo$Cilla

# fit model
GUESS <- ctmm.guess(Cilla,interactive=FALSE)
FIT <- ctmm.fit(Cilla,GUESS)

# Tell us something interpretable
summary(FIT)

#> $name
#> [1] "OUF anisotropic"
#> 
#> $DOF
#>       mean       area      speed 
#>   10.73354   18.13597 3445.13307 
#> 
#> $CI
#>                                 low        est      high
#> area (square kilometers) 239.647400 403.458801 609.24116
#> τ[position] (days)         4.026012   7.505367  13.99165
#> τ[velocity] (minutes)     39.570325  42.069009  44.72547
#> speed (kilometers/day)    13.819957  14.054636  14.28926
#> 

# }