Summarize a continuous-time movement model

This function returns a list of biologically interesting parameters in human readable format, as derived from a continuous-time movement model.

# S3 method for ctmm
summary(object,level=0.95,level.UD=0.95,units=TRUE,IC="AICc",...)

Arguments

object
A ctmm movement-model object from the output of ctmm.fit.

level
Confidence level for parameter estimates.

level.UD
Confidence level for the Gaussian home-range area.

units
Convert result to natural units.

IC
Information criteria for sorting lists of ctmm objects. Options include "AIC", "AICc", "BIC", and "MSPE".

...
Unused options.

Value

If summary is called with a single ctmm object output from ctmm.fit, then a list is returned with effective sample size array DOF and parameter estimate table CI, with low, maximum likelihood, and high estimates for the following possible parameters:

tau

The autocorrelation timescales.

area

The Gaussian home-range area, where the point estimate has a significance level of level.UD. I.e., the core home range is where the animal is located 50% of the time with level.UD=0.50. This point estimate itself is subject to uncertainty, and is given confidence intervals derived from level.

speed

The Gaussian root-mean-square (RMS) velocity, which is a convenient measure of average speed but not the conventional measure of average speed (see speed).

If summary is called on a list of ctmm objects output from ctmm.select, then a table is returned with the model names and AIC differences, where "IID" denotes the uncorrelated bi-variate Gaussian model, "OU" denotes the continuous-position Ornstein-Uhlenbeck model, and "OUF" denotes the continuous-velocity Ornstein-Uhlenbeck-F model.

Note

Confidence intervals on the autocorrelation timescales assume they are sufficiently greater than zero and less than infinity.

In ctmm v0.3.4 the speed estimate was fixed to be the RMS velocity and not \(1/\sqrt{2}\) times the RMS velocity.

See also

Examples

# Load package and data
library(ctmm)
data(buffalo)

# Extract movement data for a single animal
Cilla <- buffalo$Cilla

# fit model
GUESS <- ctmm.guess(Cilla,interactive=FALSE)
FIT <- ctmm.fit(Cilla,GUESS)

# Tell us something interpretable
summary(FIT)

#> $DOF
#>     mean     area 
#> 11.54854 22.49909 
#> 
#> $CI
#>                                 low        ML      high
#> area (square kilometers) 234.954080 372.73449 541.79530
#> tau position (days)        4.039193   6.92318  11.86634
#> tau velocity (minutes)    39.636583  42.14108  44.80384
#> speed (kilometers/day)    13.818785  14.05342  14.28799
#>