Summarize a continuous-time movement model

This function returns a list of biologically interesting parameters in human readable format, as derived from a continuous-time movement model.

# S3 method for ctmm
summary(object,level=0.95,level.UD=0.95,units=TRUE,IC=NULL,MSPE=NULL,...)

Arguments

object

A ctmm movement-model object from the output of ctmm.fit.

level

Confidence level for parameter estimates.

level.UD

Coverage level for the Gaussian home-range area.

units

Convert result to natural units.

IC

Information criteria for sorting lists of ctmm objects. Can be "AICc", "AIC", "BIC", "LOOCV", "HSCV", or none (NA). AICc is approximate.

MSPE

Sort models with the same autocovariance structure by the mean square predictive error of "position", "velocity", or not (NA).

...

Unused options.

Value

If summary is called with a single ctmm object output from ctmm.fit, then a list is returned with the effective sample sizes of various parameter estimates (DOF) and a parameter estimate table CI, with low, point, and high estimates for the following possible parameters:

tau

The autocorrelation timescales. tau position is also the home-range crossing timescale.

area

The Gaussian home-range area, where the point estimate has a significance level of level.UD. I.e., the core home range is where the animal is located 50% of the time with level.UD=0.50. This point estimate itself is subject to uncertainty, and is given confidence intervals derived from level.

This Gaussian estimate differs from the kernel density estimate of summary.UD. The Gaussian estimate has more statistical efficiency, but is less related to space use for non-Gaussian processes.

speed

The Gaussian root-mean-square (RMS) velocity, which is a convenient measure of average speed but not the conventional measure of average speed (see speed).

If summary is called on a list of ctmm objects output from ctmm.select, then a table is returned with the model names and IC differences for comparison across autocovariance structures. The mean square prediction error (MSPE) is also returned for comparison across trend structures (with autocovariance structure fixed). For the model names, "IID" denotes the uncorrelated bi-variate Gaussian model, "OU" denotes the continuous-position Ornstein-Uhlenbeck model, "OUF" denotes the continuous-velocity Ornstein-Uhlenbeck-F model, "OUf" denotes the OUF model where the two autocorrelation timescales cannot be statistically distinguished.

Author

C. H. Fleming.

Note

Confidence intervals on the autocorrelation timescales assume they are sufficiently greater than zero and less than infinity.

IC="LOOCV" can only be attempted if also specified during ctmm.select, as this argument requires additional calculations.

Prior to ctmm v0.6.2, timescale confidence intervals were constructed from normal and inverse-normal sampling distributions, whereas v0.6.2 onward uses gamma and inverse-gamma sampling distributions.

In ctmm v0.5.1 onward the MSPE is averaged over all possible times instead of over all sampled times.

In ctmm v0.3.4 the speed estimate was fixed to be the RMS velocity and not \(1/\sqrt{2}\) times the RMS velocity.

See also

Examples

# \donttest{
# Load package and data
library(ctmm)
data(buffalo)

# Extract movement data for a single animal
DATA <- buffalo$Cilla

# fit model
GUESS <- ctmm.guess(DATA,interactive=FALSE)
FIT <- ctmm.fit(DATA,GUESS)

# Tell us something interpretable
summary(FIT)
#> $name
#> [1] "OUF anisotropic"
#> 
#> $DOF
#>       mean       area  diffusion      speed 
#>   10.73354   18.13601  902.23377 3445.13306 
#> 
#> $CI
#>                                          low        est       high
#> area (square kilometers)          239.647411 403.458581 609.240606
#> τ[position] (days)                  4.438956   7.505363  12.690025
#> τ[velocity] (minutes)              39.607116  42.069009  44.683929
#> speed (kilometers/day)             13.820458  14.055146  14.289780
#> diffusion (square kilometers/day)   5.284545   5.647059   6.021428
#> 
# }